the bandwith is measured in bps, no need for decimal values
git-svn-id: https://xpra.org/svn/Xpra/trunk@21240 3bb7dfac-3a0b-4e04-842a-767bc560f471

+++ src/xpra/server/window/window_stats.py
@@ -274,7 +274,7 @@ class WindowPerformanceStatistics(object):
         elapsed = recs[-1][0]-recs[0][0]
         if elapsed==0:
             return 0
-        return bits/elapsed
+        return int(bits/elapsed)
 
     def get_damage_pixels(self, elapsed=1):
         cutoff = monotonic_time()-elapsed
